Moved the volunteer opportunity relationship tests to the test opportunity service file
